ja zoek het uti

+++ Walgelijk.OpenTK.MotionTK/MotionTK/AudioPacket.cs
@@ -2,18 +2,18 @@
 {
     public unsafe class AudioPacket : Packet
     {
-        public readonly short[] SampleBuffer;
+        public readonly byte[] SampleBuffer;
         public readonly int TotalSampleCount;
 
         public AudioPacket(byte* sampleBuffer, int sampleCount, int channelCount)
         {
-            TotalSampleCount = sampleCount * channelCount ;
-            SampleBuffer = new short[TotalSampleCount];
+            TotalSampleCount = sampleCount * channelCount;
+            SampleBuffer = new byte[TotalSampleCount];
 
             // copy buffer
             for (int i = 0; i < TotalSampleCount; i++)
             {
-                SampleBuffer[i] = ((short*)sampleBuffer)[i];
+                SampleBuffer[i] = ((byte*)sampleBuffer)[i];
             }
         }
     }
